Test for Repositories

+++ src/main/java/com/javacourse/bookstore/controllers/AuthorController.java
@@ -5,6 +5,7 @@ import com.javacourse.bookstore.domain.dto.AuthorRespDTOStock;
 import com.javacourse.bookstore.domain.dto.AuthorRespDTOWithBooks;
 import com.javacourse.bookstore.services.AuthorServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -20,36 +21,66 @@ public class AuthorController {
     }
 
     @GetMapping("/authors/{id}")
-    public ResponseEntity<AuthorRespDTOStock> getBookByID(@PathVariable("id") long id) {
-        return ResponseEntity.status(200).body(authorServiceImpl.getAuthorByID(id));
+    public ResponseEntity<? super AuthorRespDTOStock> getBookByID(@PathVariable("id") long id) {
+        AuthorRespDTOStock authorByID = authorServiceImpl.getAuthorByID(id);
+        if (authorByID != null) {
+            return ResponseEntity.status(200).body(authorByID);
+        } else {
+            return new ResponseEntity<>("Author not found ", HttpStatus.NOT_FOUND);
+        }
     }
+
     @GetMapping("/authorBooks/{idBook}")
-    public ResponseEntity<AuthorRespDTOStock> getAuthorByBook(@PathVariable("idBook") long idBook) {
-        return ResponseEntity.status(200).body(authorServiceImpl.findAuthorByBook(idBook));
+    public ResponseEntity<? super AuthorRespDTOStock> getAuthorByBook(@PathVariable("idBook") long idBook) {
+        AuthorRespDTOStock authorByBook = authorServiceImpl.findAuthorByBook(idBook);
+        if (authorByBook != null) {
+            return ResponseEntity.status(200).body(authorByBook);
+        } else {
+            return new ResponseEntity<>("Author not found ", HttpStatus.NOT_FOUND);
+        }
     }
 
     @GetMapping("/authors")
-    public ResponseEntity<List<AuthorRespDTOStock>> allBooks() {
-        return ResponseEntity.status(200).body(authorServiceImpl.getAllAuthor());
+    public ResponseEntity<? super List<AuthorRespDTOStock>> allAuthors() {
+        List<AuthorRespDTOStock> allAuthor = authorServiceImpl.getAllAuthor();
+        if (allAuthor != null) {
+            return ResponseEntity.status(200).body(allAuthor);
+        } else {
+            return new ResponseEntity<>("Authors not found ", HttpStatus.NOT_FOUND);
+        }
     }
+
     @GetMapping("/authorWith/{idAuthor}")
-    public ResponseEntity<AuthorRespDTOWithBooks> getAuthorWithDetails(@PathVariable("idAuthor") long idAuthor) {
-        return ResponseEntity.status(200).body(authorServiceImpl.getAuthorWithDetails(idAuthor));
+    public ResponseEntity<? super AuthorRespDTOWithBooks> getAuthorWithDetails(@PathVariable("idAuthor") long idAuthor) {
+        AuthorRespDTOWithBooks authorWithDetails = authorServiceImpl.getAuthorWithDetails(idAuthor);
+        if (authorWithDetails != null) {
+            return ResponseEntity.status(200).body(authorWithDetails);
+        }
+        return new ResponseEntity<>("Authors not found ", HttpStatus.NOT_FOUND);
     }
 
     @PostMapping("/authors")
-    public ResponseEntity<AuthorRespDTOStock> create(@RequestBody(required = false) AuthorReqDTO authorReqDTO) {
-        return ResponseEntity.status(201).body(authorServiceImpl.createAuthor(authorReqDTO));
+    public ResponseEntity<? super AuthorRespDTOStock> create(@RequestBody(required = false) AuthorReqDTO authorReqDTO) {
+        if (authorReqDTO != null) {
+            ResponseEntity.status(201).body(authorServiceImpl.createAuthor(authorReqDTO));
+        }
+        return new ResponseEntity<>("Authors not found ", HttpStatus.NOT_FOUND);
 
     }
 
     @PutMapping("/authors/{id}")
-    public ResponseEntity<AuthorRespDTOStock> update(@PathVariable("id") long id, @RequestBody(required = false) AuthorReqDTO authorReqDTO) {
-        return ResponseEntity.status(200).body(authorServiceImpl.updateAuthor(id, authorReqDTO));
+    public ResponseEntity<? super AuthorRespDTOStock> update(@PathVariable("id") Long id, @RequestBody(required = false) AuthorReqDTO authorReqDTO) {
+        if (authorReqDTO != null && id != null) {
+            return ResponseEntity.status(200).body(authorServiceImpl.updateAuthor(id, authorReqDTO));
+        }
+        return new ResponseEntity<>("Authors not found ", HttpStatus.NOT_FOUND);
     }
 
     @DeleteMapping("/authors/{id}")
-    public ResponseEntity<AuthorRespDTOStock> delete(@PathVariable("id") long id) {
-        return ResponseEntity.status(200).body(authorServiceImpl.deleteAuthor(id));
+    public ResponseEntity<? super AuthorRespDTOStock> delete(@PathVariable("id") Long id) {
+        if (id != null) {
+            return ResponseEntity.status(200).body(authorServiceImpl.deleteAuthor(id));
+        }
+        return new ResponseEntity<>("Authors not found ", HttpStatus.NOT_FOUND);
     }
 }
