+++ src/main/java/com/javacourse/bookstore/controllers/BookController.java
@@ -4,6 +4,7 @@ import com.javacourse.bookstore.domain.dto.BookReqDTO;
 import com.javacourse.bookstore.domain.dto.BookRespDTO;
 import com.javacourse.bookstore.services.BookServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -19,12 +20,19 @@ public class BookController {
     }
 
     @GetMapping("/books/{id}")
-    public ResponseEntity<BookRespDTO> getBookByID(@PathVariable("id") long id) {
-        return ResponseEntity.status(200).body(bookServiceImpl.getBookById(id));
+    public ResponseEntity<? super BookRespDTO> getBookByID(@PathVariable("id") Long id) {
+        if (id != null) {
+            return ResponseEntity.status(200).body(bookServiceImpl.getBookById(id));
+        }
+        return new ResponseEntity<>("Book not found ", HttpStatus.NOT_FOUND);
     }
+
     @GetMapping("/booksAuthor/{authorID}")
-    public ResponseEntity<List<BookRespDTO>> findAllByAuthorID(@PathVariable("authorID") long authorID) {
-        return ResponseEntity.status(200).body(bookServiceImpl.allBooksAuthor(authorID));
+    public ResponseEntity<? super List<BookRespDTO>> findAllByAuthorID(@PathVariable("authorID") Long authorID) {
+        if (authorID != null) {
+            return ResponseEntity.status(200).body(bookServiceImpl.allBooksAuthor(authorID));
+        }
+        return new ResponseEntity<>("Book not found ", HttpStatus.NOT_FOUND);
     }
 
 
@@ -34,17 +42,27 @@ public class BookController {
     }
 
     @PostMapping("/books")
-    public ResponseEntity<BookRespDTO> create(@RequestBody(required = false) BookReqDTO bookReqDTO) {
+    public ResponseEntity<? super BookRespDTO> create(@RequestBody(required = false) BookReqDTO bookReqDTO) {
+        if (bookReqDTO != null) {
             return ResponseEntity.status(201).body(bookServiceImpl.create(bookReqDTO));
+        }
+        return new ResponseEntity<>("Book not found ", HttpStatus.NOT_FOUND);
     }
 
     @PutMapping("/books/{id}")
-    public ResponseEntity<BookRespDTO> update(@PathVariable("id") long id, @RequestBody(required = false) BookReqDTO bookReqDTO) {
-        return ResponseEntity.status(200).body(bookServiceImpl.update(id, bookReqDTO));
+    public ResponseEntity<? super BookRespDTO> update(@PathVariable("id") Long id, @RequestBody(required = false) BookReqDTO bookReqDTO) {
+        if (id != null && bookReqDTO != null) {
+            return ResponseEntity.status(200).body(bookServiceImpl.update(id, bookReqDTO));
+        }
+        return new ResponseEntity<>("Book not found ", HttpStatus.NOT_FOUND);
     }
 
     @DeleteMapping("/books/{id}")
-    public ResponseEntity<BookRespDTO> delete(@PathVariable("id") long id) {
-        return ResponseEntity.status(200).body(bookServiceImpl.delete(id));
+    public ResponseEntity<? super BookRespDTO> delete(@PathVariable("id") Long id) {
+        if (id != null) {
+            return ResponseEntity.status(200).body(bookServiceImpl.delete(id));
+        }
+        return new ResponseEntity<>("Book not found ", HttpStatus.NOT_FOUND);
     }
+
 }
